2/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
emoji==2.2.0
idna==3.4
pyTelegramBotAPI==4.8.0
requests==2.28.1
urllib3==1.26.13